flask_server/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
urllib3==1.25.8
Werkzeug==1.0.1
python-geohash==0.8.5
scrypt==0.8.15